+++ app/controllers/users/passwords_controller.rb
@@ -1,7 +1,16 @@
 module Users
   class PasswordsController < Devise::PasswordsController
+    include Users::Concerns::Recaptcha
+    prepend_before_action :check_captcha, only: :create
+
     after_action only: :update do
       User::GetReal.call(current_user)
+    end
+
+    private
+
+    def user_params
+      params.require(:user).permit(:email)
     end
   end
 end
