+++ app/controllers/users/registrations_controller.rb
@@ -1,8 +1,8 @@
 module Users
   class RegistrationsController < Devise::RegistrationsController
-    attr_reader :recaptcha_error_message
+    include Users::Concerns::Recaptcha
     prepend_before_action :check_captcha, only: :create
-    helper_method :recaptcha_error_message
+    alias_method :user_params, :sign_up_params
 
     def create
       synthetic_user = User::CheckSynthetic.call(params[:user][:email])
@@ -35,20 +35,6 @@ module Users
         :password, :password_confirmation,
         :subscribed
       )
-    end
-
-    def check_captcha
-      return if verify_recaptcha
-
-      self.resource = resource_class.new sign_up_params
-      resource.validate # Look for any other validation errors besides Recaptcha
-      add_recaptcha_error
-      respond_with_navigational(resource) { render :new }
-    end
-
-    def add_recaptcha_error
-      @recaptcha_error_message = flash[:recaptcha_error]
-      flash.delete(:recaptcha_error)
     end
   end
 end
